Opening this for a short discussion. (To be clear, the questionable changes are the ones I wrote in this PR, not referring to any other code.)

The first commit moves away some APIs that were in the error module. I’m not sure if anyone actually uses these or not. It seems like they may have been intended entirely for the derive. @jam1garner do you remember anything about these? Is there a reason why they should not be moved/removed like this?

The second commit is about the fake specialisation mechanism added for `Vec<u8>`. I only noticed this problem because after switching to binrw I had a new compiler error because a generic type I’d created which accepted a `BinRead` was not also restricted to the `'static` lifetime. It’s possible to reduce the scope of the lifetime requirement by applying it only to the `Vec<B>` implementation, which is what this commit does, but I’m not sure if this is actually better or not. It does follow some principle of least constraint, but ideally there would not need to be any static lifetime restriction here at all, and having it only on `Vec<B>` seems like it could be surprising. I’m not sure if there is any safe way of avoiding it since `Any` has the `'static` requirement so that downcast call can’t happen otherwise. Could you maybe start by sharing some actual performance benchmarks for this fake specialisation (when using a BufRead)? That would help me to decide where I land on its existence.