@jam1garner 2021-03-23
Those APIs were intended to be macro-only and private, correct.

Benchmark:
```rust
use binrw::{BinRead, BinReaderExt, NullWideString};
use std::io::BufReader;

fn main() {
    #[derive(BinRead)]
    struct Test {
        #[br(count = 0x400_000)]
        bytes: Vec<u8>,
    }

    let mut x = BufReader::new(std::fs::File::open("/dev/zero").unwrap());
    let y: Test = x.read_be().unwrap();
}
```

Result without specialization:
```
  Time (mean ± σ):      2.193 s ±  0.041 s    [User: 253.7 ms, System: 1935.5 ms]
  Range (min … max):    2.142 s …  2.270 s    10 runs
```

Result with:
```
  Time (mean ± σ):       3.8 ms ±   0.6 ms    [User: 1.1 ms, System: 2.7 ms]
  Range (min … max):     1.5 ms …   5.6 ms    625 runs
```

System specs:
![image](https://user-images.githubusercontent.com/8260240/112227090-732d3600-8c05-11eb-9dee-a5025d9615fb.png)