@csnover 2021-04-09
Thank you for the benchmark, that is very good (/disappointing :-)). After letting this sit in my brain for a while I think that the `'static` restriction is fine since, at worst, there is a clear path to being able to eliminate it later whenever specialisation finally makes it into the language and the Any-hack goes away. Since the other half of this PR was already committed, I will close this.